arm_vcpu
Type Alias TrapFrame

pub type TrapFrame = Aarch64ContextFrame;

context frame for aarch64

Aliased Type

struct TrapFrame {
    pub gpr: [u64; 31],
    pub sp_el0: u64,
    pub elr: u64,
    pub spsr: u64,
}

Fields
gpr: [u64; 31]

An array of 31 u64 values representing the general-purpose registers.

sp_el0: u64

The stack pointer associated with EL0 (SP_EL0)

elr: u64

The exception link register, which stores the return address after an exception.

spsr: u64

The saved program status register, which holds the state of the program at the time of an exception.