Crate arm_vcpu

arm_vcpu

Definition of the vCPU structure and virtualization-related interface support for the AArch64 architecture.

Structs

Aarch64PerCpu

Per-CPU data. A pointer to this struct is loaded into TP when a CPU starts. This structure

Aarch64VCpu

A virtual CPU within a guest

Functions

has_hardware_support

Return if current platform support virtualization extension.

Type Aliases

AxArchVCpuConfig

Indicates the parameter type used for creating a vCPU, currently using VmCpuRegisters directly.

TrapFrame

context frame for aarch64